Roll tests/conftest into .testing; restore coverage
- Use importlib.resources to identify path to test data.
