+++ genno/testing/__init__.py
@@ -1,4 +1,5 @@
 import contextlib
+import importlib.resources
 import logging
 from copy import copy
 from functools import partial
@@ -18,6 +19,8 @@ from genno import ComputationError, Computer, Key, Quantity
 from genno.core.sparsedataarray import HAS_SPARSE
 
 log = logging.getLogger(__name__)
+
+# Pytest hooks
 
 
 def pytest_runtest_makereport(item, call):
@@ -371,30 +374,6 @@ def assert_units(qty: Quantity, exp: str) -> None:
     ).dimensionless, f"Units '{qty.units:~}'; expected {repr(exp)}"
 
 
-@pytest.fixture(params=[(True, "AttrSeries"), (HAS_SPARSE, "SparseDataArray")])
-def parametrize_quantity_class(request):
-    """Fixture to run tests twice, for both Quantity implementations."""
-    if not request.param[0]:
-        pytest.skip(reason="`sparse` not available → can't test SparseDataArray")
-
-    pre = genno.core.quantity.CLASS
-
-    genno.core.quantity.CLASS = request.param[1]
-    yield
-
-    genno.core.quantity.CLASS = pre
-
-
-@pytest.fixture(scope="function")
-def quantity_is_sparsedataarray(request):
-    pre = copy(genno.core.quantity.CLASS)
-
-    genno.core.quantity.CLASS = "SparseDataArray"
-    yield
-
-    genno.core.quantity.CLASS = pre
-
-
 def random_qty(shape: Dict[str, int], **kwargs):
     """Return a Quantity with `shape` and random contents.
 
@@ -422,3 +401,51 @@ def random_qty(shape: Dict[str, int], **kwargs):
         ),
         **kwargs,
     )
+
+
+# Fixtures
+
+
+@pytest.fixture(scope="session")
+def test_data_path():
+    """Path to the directory containing test data."""
+    return importlib.resources.files("genno.tests.data")
+
+
+@pytest.fixture(scope="session")
+def ureg():
+    """Application-wide units registry."""
+    registry = pint.get_application_registry()
+
+    # Used by .compat.ixmp, .compat.pyam
+    for name in ("USD", "case"):
+        try:
+            registry.define(f"{name} = [{name}]")
+        except pint.RedefinitionError:  # pragma: no cover
+            pass
+
+    yield registry
+
+
+@pytest.fixture(params=[(True, "AttrSeries"), (HAS_SPARSE, "SparseDataArray")])
+def parametrize_quantity_class(request):
+    """Fixture to run tests twice, for both Quantity implementations."""
+    if not request.param[0]:
+        pytest.skip(reason="`sparse` not available → can't test SparseDataArray")
+
+    pre = genno.core.quantity.CLASS
+
+    genno.core.quantity.CLASS = request.param[1]
+    yield
+
+    genno.core.quantity.CLASS = pre
+
+
+@pytest.fixture(scope="function")
+def quantity_is_sparsedataarray(request):
+    pre = copy(genno.core.quantity.CLASS)
+
+    genno.core.quantity.CLASS = "SparseDataArray"
+    yield
+
+    genno.core.quantity.CLASS = pre
